requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
plotly>=5.17.0
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
playwright>=1.40.0
xhtml2pdf>=0.2.11
openpyxl>=3.1.0
pandas>=2.0.0